requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.21.0
pandas==1.3.0
yfinance==0.1.63
requests==2.26.0
tensorflow==2.8.0
scikit-learn==1.0.2
pydantic==2.6.1
openai==1.12.0
python-dotenv==1.0.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0 